Renamed files to match naming convention of files
