+++ frontend2/src/components/tables/chart/Chart.ts
