+++ frontend2/src/components/tables/chart/RandomData.ts
